Etape 2.21: terminée + réorganisation

+++ Etape 2.21/Src/principal.c
@@ -1,13 +1,12 @@
 /*Partie 2.2.1*/
 #include "gassp72.h"
 
-extern short TabCos;
-extern short TabSin;
-extern short TabSig;
-int dftint(short *, int, short *);
+extern int TabCos;
+extern int TabSin;
+extern int TabSig;
+int dftint(int *, int, int *);
 int TabReel[64];
 int TabIm[64];
-//int res = 0;
 
 
 int main(void)
